# docs/swagger.yaml
basePath: /
definitions:
  domain.Account:
    properties:
      account_id:
        type: string
      account_type:
        type: string
      balance:
        type: integer
      color_hex:
        type: string
      created_at:
        type: string
      currency:
        type: string
      external_account_id:
        type: string
      is_archived:
        type: boolean
      is_imported:
        type: boolean
      last_synced_at:
        type: string
      name_account:
        type: string
      user_id:
        type: string
    type: object
...
  handler.CreateAccountReq:
    properties:
      account_type:
        example: manual
        type: string
      color_hex:
        example: '#FF0000'
        type: string
      currency:
        example: RUB
        type: string
      external_account_id:
        example: "null"
        type: string
      initial_balance:
        example: 100000
        type: integer
      is_imported:
        example: false
        type: boolean
      name:
        example: Мой кошелек
        type: string
    type: object
...
  handler.UpdateAccountReq:
    properties:
      initial_balance:
        example: 150000
        type: integer
      name:
        example: Новое название кошелька
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для приложения по управлению личными финансами.
  title: Finance Manager API
  version: "1.0"
paths:
  /api/v1/accounts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Account'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Получить все активные счета
      tags:
      - accounts
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные счета
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAccountReq'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать счет
      tags:
      - accounts
  /api/v1/accounts/{id}:
    delete:
      parameters:
      - description: ID счета
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Архивировать (удалить) счет
      tags:
      - accounts
    put:
      consumes:
      - application/json
      parameters:
      - description: ID счета
        in: path
        name: id
        required: true
        type: string
      - description: Название и/или начальный баланс
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAccountReq'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить ручной счет
      tags:
      - accounts
  /api/v1/accounts/{id}/sync/pdf:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: ID счета
        in: path
        name: id
        required: true
        type: string
      - description: PDF выписка Т-Банка
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Синхронизировать импортированный счет по PDF выписке
      tags:
      - accounts
  /api/v1/accounts/import/pdf:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: PDF выписка Т-Банка
        in: formData
        name: file
        required: true
        type: file
      - description: Название счета
        in: formData
        name: name
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Импортировать счет и транзакции из PDF выписки Т-Банка
      tags:
      - accounts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"